Clarify requirements.txt
The line in the README about requirements.txt was a little sparse, so I added more explanation in the file itself, Gemfile-style. Commented out requirements, so that users have to un-comment requirements based on their use case. Also explains why we don't need to require the PostgreSQL driver here, and why NOT to comment out simple-crypt.